Split action logic from object file
